Build fix: Get rid of second call to max
By changing the type of the member variable _timeout_count to uint32_t
the likelyhood of this wrapping around is minimized so we can remove
the second call to max that failed the build.

So lets remove the call to max in order to be able to compile this
great library. As a side note the call to max looks like it should
have been a call to min.

+++ PingSerial.h
@@ -106,7 +106,7 @@ class PingSerial {
       unsigned long _op_started = 0;
       unsigned long _max_op_duration_ms = 0;
       uint16_t      _op_timeout_ms = 0;
-      uint16_t      _timeout_count = 0;
+      uint32_t      _timeout_count = 0;
 
       // Stored values
       uint16_t _distance = 0; // Distance is always positive, max (255 * 256 + 255) = 65535
